# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: آل ياسين لاكسسوار الموتال - نظام الكاشير
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierBody"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseBody"
      responses:
        "201":
          description: Created purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "404":
          description: Not found
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales/invoices
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: customerName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale/invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Not found
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalDebt:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - totalDebt
        - createdAt
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
      required:
        - name
    UpdateSupplierBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    Purchase:
      type: object
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        invoiceNumber:
          type: ["string", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        paymentType:
          type: string
          enum: [cash, credit, partial]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        createdAt:
          type: string
      required:
        - id
        - supplierId
        - supplierName
        - totalAmount
        - paidAmount
        - remainingAmount
        - paymentType
        - items
        - createdAt
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        costPrice:
          type: number
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - costPrice
        - total
    CreatePurchaseBody:
      type: object
      properties:
        supplierId:
          type: integer
        invoiceNumber:
          type: string
        paymentType:
          type: string
          enum: [cash, credit, partial]
        paidAmount:
          type: number
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreatePurchaseItemBody"
      required:
        - supplierId
        - paymentType
        - paidAmount
        - items
    CreatePurchaseItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        costPrice:
          type: number
      required:
        - productId
        - quantity
        - costPrice
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        totalAmount:
          type: number
        paidAmount:
          type: number
        discountAmount:
          type: number
        previousDebt:
          type: number
        remainingDebt:
          type: number
        paymentType:
          type: string
          enum: [cash, credit, partial]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
        cashierName:
          type: string
      required:
        - id
        - invoiceNumber
        - totalAmount
        - paidAmount
        - discountAmount
        - previousDebt
        - remainingDebt
        - paymentType
        - items
        - createdAt
        - cashierName
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        sellingPrice:
          type: number
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - sellingPrice
        - total
    CreateSaleBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        paymentType:
          type: string
          enum: [cash, credit, partial]
        paidAmount:
          type: number
        discountAmount:
          type: number
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleItemBody"
      required:
        - paymentType
        - paidAmount
        - discountAmount
        - items
    CreateSaleItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        sellingPrice:
          type: number
      required:
        - productId
        - quantity
        - sellingPrice
...