# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * آل ياسين لاكسسوار الموتال - نظام الكاشير
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSaleBodyPaymentType } from "./createSaleBodyPaymentType";
import type { CreateSaleItemBody } from "./createSaleItemBody";

export interface CreateSaleBody {
  customerName?: string;
  customerPhone?: string;
  paymentType: CreateSaleBodyPaymentType;
  paidAmount: number;
  discountAmount: number;
  notes?: string;
  items: CreateSaleItemBody[];
}